Optimization: VPK line re-ordering for size
This is a really weird commit that probably won't have been worth the
effort (although it I did use it as an opportunity to comment more
things.) I went through all of the code and re-ordered lines in every
function where possible to reduce the size of the VPK file. This does
not always make sense and doesn't directly correlate with the bin size.
It's so weird that removing things from the code can *increase* the file
size. I was able to shave off about 400 bytes total with that and some
logic simplification. But it is probably very easy to lose this savings
when adding more features.